Final healpix maps have been written to file

+++ mapsBuilder/include/extractor.h
@@ -16,6 +16,6 @@ extern void extract_from_distribution(
     const std::vector<double> old_pointing,
     const std::vector<double> pointing,
     const TRandom3 &rgen,
-    std::vector<std::vector<unsigned int>> &pixel_dataMap);
+    std::vector<std::vector<float>> &pixel_dataMap);
 
 #endif
